test(core): match tar parser fixtures and more real world examples
